Rename to `github-actions-indicators`
